+++ CHANGELOG.md
@@ -1,13 +1,17 @@
 # Javis.jl - Changelog
 
-## Unreleased (v0.5.4)
+## Unreleased
 - added support for `rescale_factor` keyword in `render` function
+
+## v0.6.0 (3rd of August 2021)
+- Added layers see `@JLayer`
 
 ## v0.5.3 (26th of July 2021)
 - Allow all kinds of iterable ways in the `act!` function such that `act!(::Matrix, ::Action)` also works
 - Updated `anim_translate`
   - Docstring: `anim_translate` translates by a vector instead of to a point
   - from->to assumes that we are at `from` already instead of adding it to it
+- Morphing mutates the object function 
 
 ## v0.5.2
 - added support for local network live streaming